alsa: alsa_object_write should handle EBADFD EPIPE and short write.
It's better to just retry the failed write after -EBADFD,
instead of trying to check the state of the pcm and calling
snd_pcm_prepare to fix it before doing the initial write.
The latter is racy, as the state of the pcm could change between
the return from snd_pcm_state() and the call to snd_pcm_writei().

Also we now account for short writes, where the count of
frames actually written is less than the number that should
be written.  (Happens in case of signal or xrun
according to alsa docs).